requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.11.12
coverage==7.13.0
frozenlist==1.8.0
idna==3.11
iniconfig==2.3.0
magic-filter==1.0.12
multidict==6.7.0
packaging==25.0
pluggy==1.6.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dotenv==1.2.1
redis==7.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
